% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_column}
\alias{add_column}
\title{Add a column definition}
\usage{
add_column(
  table_id,
  column_definition,
  before = FALSE,
  position = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{column_definition}{The column definition. The output of the
\code{tabulator_column} function}

\item{before}{Whether the column should be added before or after}

\item{position}{The position of the column}

\item{session}{The Shiny session}
}
\description{
Add a column definition
}
