% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{add_sheet}
\alias{add_sheet}
\title{Add a sheet to a spreadsheet}
\usage{
add_sheet(
  table_id,
  title,
  key,
  session = shiny::getDefaultReactiveDomain(),
  ...
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{title}{The title of the sheet}

\item{key}{The key of the sheet}

\item{session}{The session to send the custom message to. Defaults to the default shiny session.}

\item{...}{Additional arguments to be passed to the \code{sheet} function.}
}
\description{
Add a sheet to a spreadsheet
}
