% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{check_valid_event_name}
\alias{check_valid_event_name}
\title{Checks for valid event names and returns a list of valid event names}
\usage{
check_valid_event_name(event_name)
}
\arguments{
\item{event_name}{A list of event names to be checked}
}
\value{
List of valid event names
}
\description{
Checks for valid event names and returns a list of valid event names
}
