% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{clear_sheet}
\alias{clear_sheet}
\title{Clear a sheet from a spreadsheet
This function clears a sheet from a spreadsheet.}
\usage{
clear_sheet(table_id, key, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}

\item{key}{The key of the sheet to clear}

\item{session}{The session to send the custom message to. Defaults to the default shiny session.}
}
\description{
Clear a sheet from a spreadsheet
This function clears a sheet from a spreadsheet.
}
