% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{column_layout_mode}
\alias{column_layout_mode}
\title{Set the column layout mode}
\usage{
column_layout_mode(tabulator_object, mode)
}
\arguments{
\item{tabulator_object}{An object of class tabulator}

\item{mode}{A character}
}
\value{
An object of class tabulator
}
\description{
You can choose how your table should layout its columns by setting the
layout mode.
}
\seealso{
\href{https://tabulator.info/docs/6.2/layout#layout}{tabulator documentation}

\code{get_available_column_layout_modes()} for a list of the available
modes.
}
