% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{copy_table}
\alias{copy_table}
\title{Copy the table to the clipboard}
\usage{
copy_table(table_id, what = "all", session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}

\item{what}{Which rows to copy. Available options are
"active", "all", "selected", "visible"}
}
\description{
Copy the table to the clipboard
}
