% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{default_sql_request_handler}
\alias{default_sql_request_handler}
\title{Default Request Handler for SQL
Intended to be used with a database connection along with pagination when constructing
the table.}
\usage{
default_sql_request_handler(tbl_db)
}
\arguments{
\item{tbl_db}{A tbl object. Could be a lazy table from the dplyr package.}
}
\description{
Default Request Handler for SQL
Intended to be used with a database connection along with pagination when constructing
the table.
}
