% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{delete_column}
\alias{delete_column}
\title{Delete a column}
\usage{
delete_column(table_id, field, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}

\item{field}{The field to be deleted}

\item{session}{The Shiny session}
}
\description{
Delete a column
}
