% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{df_to_values_list}
\alias{df_to_values_list}
\title{Convert a data frame to a list of values}
\usage{
df_to_values_list(df)
}
\arguments{
\item{df}{A data frame to convert to a list of values}
}
\value{
A list of values representing the data frame
}
\description{
Convert a data frame to a list of values
}
