% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{enable_copy_paste}
\alias{enable_copy_paste}
\title{Enable copy paste for a spreadsheet}
\usage{
enable_copy_paste(tabulator_object)
}
\arguments{
\item{tabulator_object}{A tabulator object}
}
\value{
The modified tabulator object
}
\description{
Enable copy paste for a spreadsheet
}
