% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{enable_range_selection}
\alias{enable_range_selection}
\title{Enable range selection for a spreadsheet}
\usage{
enable_range_selection(tabulator_object)
}
\arguments{
\item{tabulator_object}{A tabulator object}
}
\value{
The modified tabulator object
}
\description{
This function enables range selection for a spreadsheet.
Range selection allows users to select a range of cells and
perform operations on multiple cells at once.
}
