% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{flatten_columns}
\alias{flatten_columns}
\title{Flatten nested column lists into a flat list of column objects}
\usage{
flatten_columns(columns)
}
\arguments{
\item{columns}{A list or vector that may contain nested column definitions}
}
\value{
A flat list of column objects
}
\description{
Flatten nested column lists into a flat list of column objects
}
\keyword{internal}
