% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{get_available_column_layout_modes}
\alias{get_available_column_layout_modes}
\title{Available Column Layout Modes}
\usage{
get_available_column_layout_modes()
}
\value{
A character The available options to set the column layout mode
}
\description{
Available Column Layout Modes
}
