% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{get_available_pagination_modes}
\alias{get_available_pagination_modes}
\title{Lists the available pagination modes}
\usage{
get_available_pagination_modes()
}
\description{
Lists the available pagination modes
}
