% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{get_example_builtin_param}
\alias{get_example_builtin_param}
\title{Get an example list of parameters for a given builtin formatter}
\usage{
get_example_builtin_param(builtin_formatter)
}
\arguments{
\item{builtin_formatter}{The name of the builtin formatter}
}
\description{
Get an example list of parameters for a given builtin formatter
}
