% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_total_pages}
\alias{get_total_pages}
\title{Helper function to get total number of pages}
\usage{
get_total_pages(.data, page_size = 10)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{page_size}{The number of rows per page}
}
\description{
Helper function to get total number of pages
}
