% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_valid_event_names}
\alias{get_valid_event_names}
\title{Gets the valid event names to be passed to a tabulator object}
\usage{
get_valid_event_names()
}
\value{
List of valid event names
}
\description{
Gets the valid event names to be passed to a tabulator object
}
