% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_valid_table_options}
\alias{get_valid_table_options}
\title{Gets the valid table options to be passed to a tabulator object}
\usage{
get_valid_table_options()
}
\description{
Gets the valid table options to be passed to a tabulator object
}
