% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_api.R
\name{hide_column_by_field}
\alias{hide_column_by_field}
\title{Hide a column by field name}
\usage{
hide_column_by_field(
  table_id,
  field,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{field}{The field to be shown}

\item{session}{The Shiny session}
}
\description{
Hide a column by field name
}
