% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{js}
\alias{js}
\title{tag javascript for rendering in tabulatoR}
\usage{
js(x)
}
\arguments{
\item{x}{a string that should be converted to a js function or symbol for use int he tabulator configs, etc.}
}
\description{
tag javascript for rendering in tabulatoR
}
