% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_data}
\alias{people_data}
\title{Sample Data Frame
#' A dataset containing sample data with 6 columns and 20 rows.}
\format{
A data frame with 20 rows and 6 variables:
\describe{
\item{Name}{Random names}
\item{Progress}{Progress between 1 and 100}
\item{Gender}{Gender}
\item{Rating}{Number between 1 and 5}
\item{Favourite_Color}{Favourite color}
\item{Date_Of_Birth}{Date of birth}
\item{Driver}{Boolean}
}
}
\source{
Generated using random sampling
}
\usage{
people_data
}
\description{
Sample Data Frame
#' A dataset containing sample data with 6 columns and 20 rows.
}
\keyword{datasets}
