% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_static}
\alias{preview_static}
\title{Preview a static Tabulator table}
\usage{
preview_static()
}
\value{
A \code{shiny.appobj} object.
}
\description{
Launches a minimal Shiny application showing a non-editable Tabulator table.
This can be used to quickly preview the package output or within automated tests.
}
\examples{
if (interactive()) {
    preview_static()
}
}
