% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo_to_chr.R
\name{quo_to_chr}
\alias{quo_to_chr}
\title{Quosure to vector of strings}
\usage{
quo_to_chr(x)
}
\arguments{
\item{x}{A quosure (\code{rlang::quos()})}
}
\description{
Designed for functions that want a vector of strings as the variable name input
}
