% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_api.R
\name{redo_action}
\alias{redo_action}
\title{Redo an action for the given table}
\usage{
redo_action(table_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}
}
\description{
Redo an action for the given table
}
