% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{remove_sheet}
\alias{remove_sheet}
\title{Remove a sheet from a spreadsheet}
\usage{
remove_sheet(table_id, key, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}

\item{key}{The key of the sheet to remove}

\item{session}{The session to send the custom message to. Defaults to the default shiny session.}
}
\description{
Remove a sheet from a spreadsheet
}
