% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_api.R
\name{replace_column_definitions}
\alias{replace_column_definitions}
\title{Replace column definitions}
\usage{
replace_column_definitions(
  table_id,
  columns,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{columns}{The column definitions}

\item{session}{The Shiny session}
}
\description{
Replace column definitions
}
