% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{set_layout_columns_on_new_data}
\alias{set_layout_columns_on_new_data}
\title{Sets layout columns on new data inserted}
\usage{
set_layout_columns_on_new_data(tabulator_object)
}
\arguments{
\item{tabulator_object}{An object of class tabulator}
}
\value{
An object of class tabulator
}
\description{
To keep the layout of the columns consistent, once the column widths have
been set on the first data load (either from the data property in the
constructor or the setData function) they will not be changed when new
data is loaded.
If you would prefer that the column widths adjust to the data each time you
load it into the table you can set the layoutColumnsOnNewData property to
true.
}
\seealso{
\href{https://tabulator.info/docs/5.5/layout#layoutcolumnsonnewdata}{tabulator documentation}
}
