% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{sheet}
\alias{sheet}
\title{Create a sheet object}
\usage{
sheet(title, key, ...)
}
\arguments{
\item{title}{A character string representing the title of the sheet}

\item{key}{A character string representing the key of the sheet (must be unique)}

\item{...}{Arbitrary named parameters}
}
\value{
A list representing a sheet object
}
\description{
Create a sheet object
}
