% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{show_column_by_field}
\alias{show_column_by_field}
\title{Show a column by field name}
\usage{
show_column_by_field(
  table_id,
  field,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{field}{The field to be shown}

\item{session}{The Shiny session}
}
\description{
Show a column by field name
}
