% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{show_example_builtin_params}
\alias{show_example_builtin_params}
\title{Show the example parameters for a given builtin formatter}
\usage{
show_example_builtin_params(builtin_formatter)
}
\arguments{
\item{builtin_formatter}{The name of the builtin formatter}
}
\description{
Show the example parameters for a given builtin formatter
}
