% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{spreadsheet}
\alias{spreadsheet}
\title{Renders the table as a spreadsheet}
\usage{
spreadsheet(tabulator_object, ...)
}
\arguments{
\item{tabulator_object}{A tabulator object}

\item{...}{Arbitrary named parameters}
}
\value{
A tabulator object with a spreadsheet enabled
}
\description{
Renders the table as a spreadsheet
}
