% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{spreadsheetClearSheet}
\alias{spreadsheetClearSheet}
\title{Clear all data from a Tabulator spreadsheet}
\usage{
spreadsheetClearSheet(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{The output ID of the Tabulator spreadsheet (string).}

\item{session}{The Shiny session object. Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\description{
Programmatically clears all data from an existing Tabulator spreadsheet in a Shiny app.
This function sends a message to the JavaScript binding to clear the spreadsheet
without re-rendering the entire widget.

This function calls Tabulator's \code{clearSheet()} method. For more information, see:
\url{https://tabulator.info/docs/6.3/spreadsheet#clear}
}
\details{
This calls Tabulator's \code{clearSheet()} method, which removes all data but preserves
the spreadsheet configuration and structure.
}
\examples{
\dontrun{
# In a Shiny server function:
observeEvent(input$clear_btn, {
  spreadsheetClearSheet("my_spreadsheet")
})
}

}
