% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{spreadsheetSetData}
\alias{spreadsheetSetData}
\title{Set data in a Tabulator spreadsheet}
\usage{
spreadsheetSetData(outputId, data, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{The output ID of the Tabulator spreadsheet (string).}

\item{data}{A data.frame, matrix, or list of lists representing the spreadsheet data.}

\item{session}{The Shiny session object. Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\description{
Programmatically loads data into an existing Tabulator spreadsheet in a Shiny app.
This function sends a message to the JavaScript binding to set the spreadsheet data
without re-rendering the entire widget.

The function automatically converts data.frames and matrices to the array-of-arrays
format required by Tabulator spreadsheets.

This function calls Tabulator's \code{setData()} method. For more information, see:
\url{https://tabulator.info/docs/6.3/data#array}
}
\details{
Unlike standard Tabulator tables, spreadsheets use an array-of-arrays data format
where each row is a list and data is accessed by position rather than field names.
}
\examples{
\dontrun{
# In a Shiny server function:
observeEvent(input$load_data_btn, {
  new_data <- data.frame(
    A = 1:10,
    B = 11:20,
    C = 21:30
  )
  spreadsheetSetData("my_spreadsheet", new_data)
})
}

}
