% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{subscribe_events}
\alias{subscribe_events}
\title{Subscribe events for the current table
Once events are subscribed, the events triggered in the browser can be listened
via registered inputs following the pattern \verb{input$<table_id>_<eventName>}. For example
if the \code{cellClick} event was subscribed for a table with an id \code{my_table} then
you can use it in your Shiny app via the input registered to \code{input$table_cellClick}
Also you can run \code{run_example_app("table_events")} to see all the events available.}
\usage{
subscribe_events(tabulator_object, events)
}
\arguments{
\item{tabulator_object}{A tabulator object}

\item{events}{List of events to subscribe. Also can be a single character}
}
\description{
Subscribe events for the current table
Once events are subscribed, the events triggered in the browser can be listened
via registered inputs following the pattern \verb{input$<table_id>_<eventName>}. For example
if the \code{cellClick} event was subscribed for a table with an id \code{my_table} then
you can use it in your Shiny app via the input registered to \code{input$table_cellClick}
Also you can run \code{run_example_app("table_events")} to see all the events available.
}
