% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Efficient tabulation}
\usage{
tab(df, ..., round = 2)
}
\arguments{
\item{df}{A data.table, tibble, or data.frame}

\item{...}{A column or set of columns (without quotation marks)}

\item{round}{An integer indicating the number of digits for proportion and cumulative proportion}
}
\value{
Tabulation (frequencies, proportion, cumulative proportion) for each unique value of the variables given in \code{...} from \code{df}.
}
\description{
Produces a tabulation: for each unique group from the variable(s),
\code{tab} shows the number of
	observations with that value, proportion of observations with that
		value, and cumulative proportion, in descending order of frequency.
		Accepts data.table or tibble as input.
Efficient with big data: if you give it a \code{data.table},
		\code{tab} uses \code{data.table} syntax.
}
\examples{
# data.table
a <- data.table(varname = sample.int(20, size = 1000000, replace = TRUE))
a \%>\% tab(varname)

# tibble
b <- tibble(varname = ample.int(20, size = 1000000, replace = TRUE))
b \%>\% tab(varname)

}
