% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tab_source}
\alias{tab_source}
\title{Generate Tabulator Source Links}
\usage{
tab_source(version = "6.3.1", theme = "tabulator_bootstrap5")
}
\arguments{
\item{version}{A character string specifying the version of Tabulator to use. Defaults to "6.3.1".}

\item{theme}{A character string specifying the theme to use. Defaults to "tabulator_bootstrap5".}
}
\value{
A \code{tagList} containing the HTML tags for the Tabulator CSS and JavaScript files.
}
\description{
This function generates the necessary HTML tags to include Tabulator's CSS and JavaScript files.
cdn for versions and css can be found at https://app.unpkg.com/tabulator-tables
}
\examples{
tab_source()
tab_source(version = "5.0", theme = "tabulator_simple")
}
