% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulatoR-package.R
\docType{package}
\name{tabulatoR-package}
\alias{tabulatoR}
\alias{tabulatoR-package}
\title{tabulatoR: Tabulator Integration for Shiny Applications}
\description{
Provides an interface for integrating Tabulator tables into Shiny applications. This package allows users to easily create interactive and editable tables with a variety of themes and customization options.
}
\author{
\strong{Maintainer}: Kent Orr \email{orr.kent@pawpawanalytics.com}

}
\keyword{internal}
