% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulator.R
\name{tabulator}
\alias{tabulator}
\title{Tabulator table}
\usage{
tabulator(
  data = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  theme = "bootstrap5"
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{width}{The width of the widget}

\item{height}{The height of the widget}

\item{elementId}{The ID of the widget}

\item{theme}{The theme of the widget}
}
\description{
Renders a tabulator table
}
