% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{tabulator_column}
\alias{tabulator_column}
\title{Tabulator column constructor
See https://tabulator.info/docs/6.2/columns#definition for more details.}
\usage{
tabulator_column(...)
}
\arguments{
\item{...}{Named arguments to be passed to the column.}
}
\description{
Tabulator column constructor
See https://tabulator.info/docs/6.2/columns#definition for more details.
}
