% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{tabulator_columns}
\alias{tabulator_columns}
\title{Set the columns for a tabulator object}
\usage{
tabulator_columns(tabulator_object, columns)
}
\arguments{
\item{tabulator_object}{A tabulator object}

\item{columns}{A list of columns}
}
\description{
Set the columns for a tabulator object
}
