% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tabulator_options}
\alias{tabulator_options}
\title{Add table level options to a tabulator object}
\usage{
tabulator_options(tabulator_object, ...)
}
\arguments{
\item{tabulator_object}{An object of class tabulator}

\item{...}{A list of options to be added as named parameters}
}
\description{
Add table level options to a tabulator object
}
\seealso{
\href{https://tabulator.info/docs/5.6/layout}{tabulator documentation}
}
