% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{tabulatorAddData}
\alias{tabulatorAddData}
\title{Append rows to a Tabulator table via proxy}
\usage{
tabulatorAddData(proxy, data, add_to = c("top", "bottom"))
}
\arguments{
\item{proxy}{A \code{tabulatorProxy} object created with \code{tabulatorProxy()}.}

\item{data}{A \code{data.frame} of rows to add.}

\item{add_to}{Whether to add rows to the "top" or "bottom" of the table.}
}
\description{
Adds new rows to the top or bottom of an existing Tabulator table.
}
