% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{tabulatorAddRow}
\alias{tabulatorAddRow}
\title{Add row(s) to a Tabulator table}
\usage{
tabulatorAddRow(
  outputId,
  data,
  add_to = "bottom",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{The output ID of the Tabulator table (string).}

\item{data}{A data.frame or list representing the row(s) to add.}

\item{add_to}{Where to add the row(s): \code{"top"} or \code{"bottom"}. Default is \code{"bottom"}.}

\item{session}{The Shiny session object. Defaults to \code{shiny::getDefaultReactiveDomain()}.}
}
\description{
Programmatically adds one or more rows to an existing Tabulator table in a Shiny app.
This function sends a message to the JavaScript binding to add rows to the table
without re-rendering the entire widget or modifying the underlying R data.
}
\details{
The \code{data} parameter can be a single-row data.frame or a multi-row data.frame.
Each row will be added to the table at the specified position.
}
\examples{
\dontrun{
# In a Shiny server function:
observeEvent(input$add_row_btn, {
  new_row <- data.frame(name = "New Person", age = 25)
  tabulatorAddRow("my_table", new_row, add_to = "top")
})
}

}
