% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tabulatoROutput}
\alias{tabulatoROutput}
\title{Create a Tabulator Output Element}
\usage{
tabulatoROutput(id, width = "100\%", height = "400px", ...)
}
\arguments{
\item{id}{A character string specifying the ID of the div element.}

\item{width}{A character string specifying the width of the div element. Defaults to "100\%".}

\item{height}{A character string specifying the height of the div element. Defaults to "400px".}

\item{...}{Additional arguments passed to \code{tab_source}.}
}
\value{
A \code{tagList} containing the HTML elements for the Tabulator output.
}
\description{
This function creates a div element for rendering a Tabulator table in a Shiny application.
}
\examples{
tabulatoROutput("myTable")
tabulatoROutput("myTable", width = "80\%", height = "500px")
}
