% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{tabulatorRestoreOldValue}
\alias{tabulatorRestoreOldValue}
\title{Restore the old value in a cell after editing}
\usage{
tabulatorRestoreOldValue(
  id,
  index,
  field,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{The ID of the Tabulator table.}

\item{index}{An integer index corresponding to the Tabulator row index.
Use the \code{index} value from an event payload (e.g., \code{input$my_table$cellClick$index}).}

\item{session}{The Shiny session object.}
}
\description{
Restores the old value in a cell after editing, useful for preserving the original if validation fails.
}
