% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{test_for_valid_table_options}
\alias{test_for_valid_table_options}
\title{Checks for valid table options and returns a list of valid options}
\usage{
test_for_valid_table_options(table_options)
}
\arguments{
\item{table_options}{A list of options to be checked}
}
\description{
Checks for valid table options and returns a list of valid options
}
