% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadsheet.R
\name{to_array_of_arrays}
\alias{to_array_of_arrays}
\title{Convert data.frame or matrix to array-of-arrays format}
\usage{
to_array_of_arrays(data)
}
\arguments{
\item{data}{A data.frame or matrix to convert}
}
\value{
A list of lists (array-of-arrays) suitable for Tabulator spreadsheet mode
}
\description{
Convert data.frame or matrix to array-of-arrays format
}
\keyword{internal}
