% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{toggle_column_by_field}
\alias{toggle_column_by_field}
\title{Toggle a column by field name}
\usage{
toggle_column_by_field(
  table_id,
  field,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{field}{The field to be shown}

\item{session}{The Shiny session}
}
\description{
Toggle a column by field name
}
