% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_api.R
\name{undo_action}
\alias{undo_action}
\title{Undo an action for the given table}
\usage{
undo_action(table_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}
}
\description{
Undo an action for the given table
}
