% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_api.R
\name{update_column_definition}
\alias{update_column_definition}
\title{Update column definition}
\usage{
update_column_definition(
  table_id,
  field,
  column_definition,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{table_id}{The ID of the table}

\item{field}{The field to be updated}

\item{column_definition}{The column definition. The output of the
\code{tabulator_column} function}

\item{session}{The Shiny session}
}
\description{
Update column definition
}
