% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{use_default_table_options}
\alias{use_default_table_options}
\title{Use default table options if they are not provided}
\usage{
use_default_table_options(tabulator_object)
}
\arguments{
\item{tabulator_object}{A tabulator object}
}
\value{
The same tabulator object but with default options added if they
were not provided
}
\description{
Use default table options if they are not provided
}
