% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history_api.R
\name{clear_history}
\alias{clear_history}
\title{Clear history for the given table}
\usage{
clear_history(table_id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{table_id}{The ID of the table}
}
\description{
Clear history for the given table
}
