% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{quantiles}
\alias{quantiles}
\title{Efficient quantiles}
\usage{
quantiles(df, ...)
}
\arguments{
\item{df}{A data.table or tibble}

\item{...}{A column or set of columns (without quotation marks)}
}
\value{
Quantile values
}
\description{
Produces quantiles of the variables.
\code{quantiles} shows quantile values.
Efficient with big data: if you give it a \code{data.table},
		\code{quantiles} uses \code{data.table} syntax.
}
\examples{
# data.table
a <- data.table(varname = sample.int(20, size = 1000000, replace = TRUE))
a \%>\% quantiles(varname)

# tibble
b <- tibble(varname = sample.int(20, size = 1000000, replace = TRUE))
b \%>\% quantiles(varname)

}
