% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{tabulatorReplaceData}
\alias{tabulatorReplaceData}
\title{Replace the data in a Tabulator table via proxy}
\usage{
tabulatorReplaceData(proxy, data)
}
\arguments{
\item{proxy}{A \code{tabulatorProxy} object created with \code{tabulatorProxy()}.}

\item{data}{A \code{data.frame} to send to the client-side Tabulator table.}
}
\description{
Replaces the current table data in the browser with a new data.frame.
This does not trigger a full re-render of the table.
}
